Add files via upload
Combines the Helios One synth from BlogHoskins and a four oscillator drone based on Control_Oscil_wash that uses the knobs and switches of the Helios. Requires one extra switch on D4 to toggle between Helios and Drone mode, and two additional 10K potentiometers on analog pins A6 and A7 for the pitch of two of the drones.
Four drone oscillators are Sine, Cos, Triangle and Saw.